fix: Fix app crash when mobile data is available but Internet access is restricted
The app is crashing upon starting if Wi-Fi is turned off and
Mobile Data is used but without having Internet access. The crash is
caused by an un-handled "StringIndexOutOfBoundsException" exception. This
exception is now handled in the try/catch clause of "getQuotes"
in the "QuoteSyncJob" class to avoid app crash on startup.

Resolves: #35